refactor(test-benchmark): centralize benchmark helper utilities (#3166)
