+++ tests/benchmark/stateful/bloatnet/test_account_query.py
@@ -28,9 +28,9 @@ from tests.benchmark.helper.account_creator import (
     AccountCreator,
     AccountMode,
 )
-from tests.benchmark.stateful.helpers import (
-    DECREMENT_COUNTER_CONDITION,
-    CacheStrategy,
+from tests.benchmark.helper.enums import CacheStrategy
+from tests.benchmark.helper.loops import DECREMENT_COUNTER_CONDITION
+from tests.benchmark.helper.transactions import (
     build_benchmark_txs,
     build_cache_strategy_blocks,
 )
