+++ tests/benchmark/stateful/bloatnet/test_call.py
@@ -18,10 +18,8 @@ from execution_testing import (
     keccak256,
 )
 
-from tests.benchmark.stateful.helpers import (
-    DECREMENT_COUNTER_CONDITION,
-    build_benchmark_txs,
-)
+from tests.benchmark.helper.loops import DECREMENT_COUNTER_CONDITION
+from tests.benchmark.helper.transactions import build_benchmark_txs
 
 
 @pytest.mark.stub_parametrize("factory_stub", "bloatnet_factory_")
