+++ tests/benchmark/stateful/bloatnet/test_transient_storage.py
@@ -12,10 +12,8 @@ from execution_testing import (
     While,
 )
 
-from tests.benchmark.stateful.helpers import (
-    DECREMENT_COUNTER_CONDITION,
-    build_benchmark_txs,
-)
+from tests.benchmark.helper.loops import DECREMENT_COUNTER_CONDITION
+from tests.benchmark.helper.transactions import build_benchmark_txs
 
 
 @pytest.mark.parametrize("with_tload", [True, False])
